requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.8
django-rest-framework==0.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
numpy==1.21.2
pandas==1.3.3
psycopg2==2.9.1
pycodestyle==2.7.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
whitenoise==5.3.0
